Fix decorrelation of WITH USING KEY (https://redirect.github.com/duckdb/duckdb/pull/16330)
Currently, the new `USING KEY` variant of recursive CTEs is not
decorrelated correctly. In addition to adding the correlated columns to
the working table, they do have to be added to the `key`.

MVP:

```sql
select  *
  from    range(1, 4) as _(l),
  lateral (
    with recursive cte(a, b) using key (a) as (
      select 1, 0
        union
      select a, b + 1
      from   recurring.cte
      where  b < l
    )
    table cte
  ) ORDER BY l
```

This query should return 3 rows, but it only returns one. This PR fixes
that.

Upstream-commit: https://github.com/duckdb/duckdb/commit/5cb575067b6f9a9544b75ca7d18cbb86fc0470b0