Verify UTF-8 in `DeltaLengthByteArrayDecoder` and speed it up (https://redirect.github.com/duckdb/duckdb/pull/16328)
Although Parquet should be valid UTF-8, we can never be sure what other
writers do, so we validate this. This validation was already there for
`PLAIN`/`RLE_DICTIONARY` encoding but was missing for
`DELTA_LENGTH_BYTE_ARRAY`. This PR adds the verification there as well.

Verifying UTF-8 takes is somewhat costly, so I've also worked on
speeding it up by checking 8 bytes at a time, instead of 1 byte. This is
especially nice for `DELTA_LENGTH_BYTE_ARRAY`, as the strings are stored
without their lengths in between, so we can verify many strings in one
go.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ef88e2bb17b1234bdc5b84c0b6ebb8e74da89bf8