Linux CLI: override platform for ARM manylinux (https://redirect.github.com/duckdb/duckdb/pull/16347)
Currently we don't distribute extensions for `linux_arm64_gcc4` flavor,
so I think it make little sense to distribute a CLI binary that can't
load them.

This is not super clean, since then there might be cases where CLI can
be launched but extensions load will fail due to a LIBC version
mismatch, but I think that's somehow better that no extensions for all.

This needs to be reviewed going into v1.3, I will add it as task.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c8a53c33073356cf7a23a00d4d1c56411c59b419