docs: tweak explanation of median for even cardinality inputs (https://redirect.github.com/duckdb/duckdb/pull/13726)
Instead of qualitative/quantitative, I think we should be more direct
and list the actual types.

Also, just a little more word-smithing to be explicit that we are only
referring to the two middle values. This change I am less convinced is a
good idea, but I still think it's a positive.

Related to https://github.com/duckdb/duckdb/issues/13655

Upstream-commit: https://github.com/duckdb/duckdb/commit/cd0d0da9b1f475632a21e11a7c00cc726bedaacc