[SwiftRelease CI] fetch tags before checking there is already a tag with the same name (https://redirect.github.com/duckdb/duckdb/pull/16376)
[SwiftRelease
CI](https://github.com/duckdb/duckdb/actions/runs/13488488943/job/37682897743#step:8:19)
fails when it tries to push with the tag name which is already exists.
The way how it checks that looks good, but maybe it should first fetch
the tag names and check only then - this PR adds a line to fetch the
tags.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0c30dce3e52a8673c84900dca42cfeccc21f9016