Rework CSV Reader: use the new MultiFileReaderFunction interface (https://redirect.github.com/duckdb/duckdb/pull/16349)
Follow-up from https://github.com/duckdb/duckdb/pull/16299

This PR reworks the CSV reader to use the new MultiFileReaderFunction.
What this means is that effectively the CSV reader no longer handles the
multi-file traversal. Instead the CSV reader implements the new
interface in the `CSVMultiFileInfo`, and the multi-file reader function
handles the opening and traversal of readers.

In order to facilitate this a few changes have been made to the CSV
reader:

* The `CSVGlobalState` has been greatly simplified - all objects related
to multi-file reading are moved to the `MultiFileReader` structures
* The `MultiFileReaderOptions` has been removed from the
`CSVReaderOptions`
* The `DecrementThreads` method has been removed. Instead, we track per
file if all tasks for that file have been completed. When a task has
been fully scanned we handle any final errors/put entries in the rejects
table.
* The CSV reader now uses the `MultiFileList` instead of a list of files

In addition, this PR reworks the way that progress is calculated during
multi-file reads. Previously we would consider any files that had all of
their scans *started* as completed. This could give skewed/odd progress
reportings when working with a few very large files. We now poll any
files that are actively being read continuously, leading to much more
smooth progress reportings. This applies to both the CSV and Parquet
reader.


### Todo

Serialization is currently disabled (although it is partially
implemented and works for most tests). There's still an open question if
this is actually used currently so we will leave it as not implemented
for now.

Upstream-commit: https://github.com/duckdb/duckdb/commit/87cf7730330dff3d1998be29915ab29bbcd6b782