Fuzzer Fix: Fix Avg for NULL cast to TIMESTAMP (https://redirect.github.com/duckdb/duckdb/pull/16394)
Fixes https://github.com/duckdb/duckdb-fuzzer/issues/3990

Fix (I think) Is to not use the DivModPositive Function as it assumes
the LHS is positive. However, it seems like when NULL is cast to
`Timestamp` a negative timestamp/hugeint_t value is used.

Another fix may be a different timestamp value when casting NULL.

Commit causing the error was
[1e404813c9ec1ebe47e4d1b4dccfd86ad978ba7d](https://github.com/duckdb/duckdb/commit/1e404813c9ec1ebe47e4d1b4dccfd86ad978ba7d).

CC @hawkfish 

Upstream-commit: https://github.com/duckdb/duckdb/commit/5cdaf5649863733392c8448f5ce4231c71e0d51d