Add a TableFilterState for execution of table filters (https://redirect.github.com/duckdb/duckdb/pull/16461)
This PR adds the `TableFilterState` which is used for the execution of
table filters. The `TableFilterState` can be used to cache thread-local
data during the evaluation of the table filter.

Currently this is not used for anything - it is mostly an
infrastructural change that is required for a follow-up change (allowing
arbitrary expressions to be evaluated as part of table filters).

Upstream-commit: https://github.com/duckdb/duckdb/commit/052b3dc20091ed4ee71f7e3218687b502bb328ea