Speed-up contains by using `memchr` on every iteration (https://redirect.github.com/duckdb/duckdb/pull/16484)
This PR reworks the implementation for `contains` (which also gets
generated as part of `LIKE '%str%'` by the optimizer).

Previously we had three different implementations - for small aligned
contains (2/4/8 bytes), small unaligned contains (3/5/6/7 bytes) and a
generic fallback for needles > 8 bytes. These implementations were
pretty distinct and not entirely optimal - they did a lot of work for
every byte even if an early-out could be used to speed this up in many
cases.

This PR unifies these three implementations and leans on `memchr` to
find the beginning of possible matches - followed by doing an actual
matching comparison. The way this works is that we first do the
comparison with the largest unsigned integer that fits in the needle (so
either 2/4/8 bytes) - followed by a `memcmp` for the remaining bytes (if
any).

Running this query on `hits` we can see this improves performance in all
cases - but especially for small unaligned searches and large searches.

```sql
SELECT COUNT(*) FROM hits WHERE URL LIKE '%goog%';
```
|     LIKE     | Search Size | v1.2.0 |  New  |
|--------------|-------------|--------|-------|
| %goog%       | 4           | 0.35s  | 0.31s |
| %google%     | 6           | 0.41s  | 0.33s |
| %google.com% | 10          | 0.43s  | 0.32s |

Upstream-commit: https://github.com/duckdb/duckdb/commit/5ca240ba2a5086a800e4fe8ee688888ac21f0068