Avoid throwing an exception (that is then swallowed) when computing compressed materialization over stats that are not set (https://redirect.github.com/duckdb/duckdb/pull/16532)
When stats are uninitialized `MIN` is set to `INT_MAX` and `MAX` is set
to `INT_MIN`, i.e. `max < min`. This PR detects this situation and
early-outs in the compressed materialization optimizer. This will
otherwise throw an exception in `ExpressionExecutor::TryEvaluateScalar`
which will get swallowed - but shows up when debugging.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c486d1797ed8f90e17d7852d200c4e1b381cfab1