Pass `ClientContext` to catalog initialize, and postpone index binding when replaying the WAL (https://redirect.github.com/duckdb/duckdb/pull/16536)
This PR passes the `ClientContext` into the `Catalog::Initialize`
method. For backwards compatibility, the old `Initialize` method is also
preserved.

Performing this change also revealed an issue in the WAL replay code -
where we would bind expressions during the `ReplayCreateIndex`. This is
solved by postponing index binding and emitting an `UnboundIndex`
instead (to be bound at a later moment).

Upstream-commit: https://github.com/duckdb/duckdb/commit/c4b52da8f6a98dd8810394f286603543f22a168b