[Dev] Make the various mappings in `MultiFileReaderData` typesafe (https://redirect.github.com/duckdb/duckdb/pull/16596)
This PR does not change anything functionality-wise

In the `MultiFileReaderData` we have many mappings, used in various
contexts, and many of these mappings are tied to each other - if one
changes, the other should also change.

All of these were using `idx_t`, which made it very confusing and
error-prone to use and understand what type of index is required and
produced.

This PR introduces the following structs as typesafe replacements:
- `MultiFileLocalColumnId`
- `MultiFileLocalIndex`
- `MultiFileGlobalIndex`
- `MultiFileConstantMapIndex`
- `MultiFileLocalColumnIds`
- `MultiFileColumnMapping`
- `MultiFileFilterMap`
- `MultiFileConstantMap`
- `MultiFileCastMap`

column_ids, column_mapping, filter_map, constant_map and cast_map are
indexed with and produce the typed indexes
`MultiFileFilterEntry` has been changed to produce
`MultiFileConstantMapIndex` or `MultiFileLocalIndex` depending on the
way it was set.

Upstream-commit: https://github.com/duckdb/duckdb/commit/81a9e52630e70ed9ea37f0c9d83a7149009ac43e