Add Notify External Repositories Workflow (https://redirect.github.com/duckdb/duckdb/pull/16730)
Currently, the ODBC driver vendors DuckDB on a nightly schedule.
However, this process does not verify whether a commit successfully
builds an extension. As a result, some ODBC versions become incompatible
with extensions due to vendoring an incomplete or incompatible commit.

This PR introduces a new workflow that allows the vendor workflow to be
triggered directly from `duckdb/duckdb`. Eventually this could be
integrated into InvokeCI, after the extension build, and then executed
on both the main and current release branch on ODBC.

Additionally, this workflow could be generalized to support other
clients by allowing input parameters for a target repository and
workflow or extended with specific runs for specific drivers.

##### Current Issues & Open Questions:
- The authentication tokens in the workflow are not functioning
correctly. Any insights on the correct tokens to use?
- Should we make this workflow more generic now or iterate on it after
validating its effectiveness for ODBC?

Would appreciate feedback on the approach and any suggestions on token
handling.

CC: @carlopi
Upstream-commit: https://github.com/duckdb/duckdb/commit/710fe9412d801812f96c72bdeb576046f3f0f260