Parquet Writer: Truncate string stats for large strings, instead of bailing on writing stats (https://redirect.github.com/duckdb/duckdb/pull/16736)
Previously we would bail on writing stats for string columns in Parquet
if they exceeded a certain size (10KB). However, this is suboptimal
because: (1) we could still write very large stats (up to 9.9KB) and (2)
we would not write any stats for large string columns.

It is not required for stats in Parquet files to actually exist in the
data. This can in newer versions be made explicit with the
`is_min_value_exact` and `is_max_value_exact` flags - which allow the
file to specify whether or not the min/max values *actually* exist in
the column or if they exist only for bounds-checking purposes.

This PR uses these flags to (almost) always emit stats for
strings/blobs. When encountering large strings, we now truncate the
min/max to 256 bytes.

The `min` value can simply be truncated. For string values, we might
truncate a bit more to ensure the string remains valid UTF8.

The `max` - the value must have one byte incremented so that it forms a
correct upper bound.

* For blob columns, we increment the first byte from the end that is not
at the maximum `0xFF` - if no such byte exists (because all 256 bytes
are `0xFF`) - we don't emit stats
* For string columns, we increment the first byte from the end that is
`< 0x7F` (i.e. within the ASCII domain). This is done so that we don't
accidentally create a string that is invalid UTF8.

Upstream-commit: https://github.com/duckdb/duckdb/commit/7c496f6dbb1f98dc6b42f4f0d597a33bdd4adec1