JSON bugfixes (https://redirect.github.com/duckdb/duckdb/pull/16729)
Fixes:
 * https://redirect.github.com/duckdb/duckdb/pull/16570
 * https://redirect.github.com/duckdb/duckdb/pull/16568
 * https://github.com/duckdblabs/duckdb-internal/issues/4389
 * https://github.com/duckdblabs/duckdb-internal/issues/4403

All of which are JSON-related. I am targeting `main`, as these bugs
aren't critical, and the JSON reader has gone through large changes on
`main`, so fixing them there saves the effort of fixing them twice.

I've also fixed https://redirect.github.com/duckdb/duckdb/pull/16572 in this PR. The issue was caused by DuckDB
preventing breadth-first plan evaluation to avoid using more memory than
necessary. However, there is no risk of running out of memory when all
of the children of a `UNION ALL` don't contain any `Sink` operators. By
checking if this is the case, we can allow more parallelism for `UNION
ALL`.

I have tested that this works with
`test/sql/parallelism/intraquery/depth_first_evaluation.test_slow`
locally (so this doesn't break the desired behavior when there are
`Sink` operators), but this test is disabled because it was causing
issues in CI.

Upstream-commit: https://github.com/duckdb/duckdb/commit/df408250eacdcbbbcb3fdf1d56a123df71fea84e