Push catalog lookups through an extensible EntryLookupInfo struct (https://redirect.github.com/duckdb/duckdb/pull/16764)
This PR reworks catalog lookups so they are performed through a struct -
`EntryLookupInfo` - which can be extended with new parameters in the
future without breaking the catalog APIs. The struct looks as follows:

```cpp
struct EntryLookupInfo {
public:
	CatalogType GetCatalogType() const;
	const string &GetEntryName() const;
	const QueryErrorContext &GetErrorContext() const;

private:
	CatalogType catalog_type;
	const string &name;
	QueryErrorContext error_context;
};
```

Several of the old retrieval methods (that do not take this struct) are
kept for backwards-compatibility reasons - but the main entry points now
all use this struct. The idea is that this will be extended in the (very
near) future with extra information that can be pushed along with the
lookup.

Upstream-commit: https://github.com/duckdb/duckdb/commit/51f1cd70bad6bec35993b4401ad7618bc4011f18