Add support for time travel syntax in the `FROM` clause (https://redirect.github.com/duckdb/duckdb/pull/16774)
This PR adds support for time travel syntax in the `FROM` clause,
specifically this syntax is now accepted:

```sql
SELECT * FROM tbl AT (TIMESTAMP => TIMESTAMP '2020-01-01')
SELECT * FROM tbl AT (VERSION => 42)
```

This is not yet supported in any catalogs, but is pushed into the
catalogs using the `EntryLookupInfo` introduced in
https://github.com/duckdb/duckdb/pull/16764. Catalogs need to indicate
they support this syntax by overloading this method, that defaults to
false:

```cpp
virtual bool SupportsTimeTravel() const {
	return false;
}
```

The information is passed to the catalog using the `BoundAtClause`:

```cpp
class BoundAtClause {
	//! The unit (e.g. TIMESTAMP or VERSION)
	string unit;
	//! The value that is associated with the unit (e.g. TIMESTAMP '2020-01-01')
	Value val;
};
```

### Keyword "AT"

In order to facilitate this in the parser, the keyword `AT` has been
moved from unreserved to partially reserved. It can no longer be used a
column name/alias without quotes.

Upstream-commit: https://github.com/duckdb/duckdb/commit/242d3f78e4651f5f99f77f33cc05d2b72a986d3d