Regression.yml: Actually checkout proper base.sha commit (https://redirect.github.com/duckdb/duckdb/pull/16824)
Previously we would blindly checkout `main`, but that might change
between when a PR is sent and when it's evaluated (due to other PR being
merged in the timeline of a PR).

Now move to use `github.event.pull_request.base.sha` for pull_requests.

Problem discovered with @c-herrewijn looking at hard to justify
behaviour on a PR of his.

Upstream-commit: https://github.com/duckdb/duckdb/commit/03432a4c56fb93e168c34d8e4da3d9fddd5d9988