[MultiFileReader] Rework `MultiFileReader::FinalizeChunk` to use Expressions (https://redirect.github.com/duckdb/duckdb/pull/16630)
Previously the underlying reader (such as csv or parquet) would read
directly into a chunk that matches the "global" schema, containing
columns that are perhaps constants for this file.
Because of this, the `column_mapping` had to be applied in the
underlying reader to find the Vector that should be written into.

This PR replaces that, letting the underlying reader be unaware of the
global schema, moving this mapping logic up to the FinalizeChunk method
of the MultiFileReader implementation.

I was hoping to remove the need to expose the `column_mapping` to the
underlying reader entirely, along with `constant_map`, `cast_map` and
`filter_map`, but this is not possible without a bigger rework.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ca2172df50a2a3a00afafc356f41be77544ea7a7