compressed vector serialization fixes (https://redirect.github.com/duckdb/duckdb/pull/16648)
binary serializer asserts tags and keys to be unique and this assert was
violated by the new compressed vector serialization

many apologies for assuming that this had been tested by
DataChunk::Verify() -- it was not because test runs still use v0.10.2 as
standard serialization compatibility.

This came out in integration testing in MotherDuck, where we have now
advanced to v1.2.0 (note that this would still not trigger the new code
as that would happen in v1.3.0 -- but in the MotherDuck cherry-pick of
this PR we have switched it off by default but can enable it with a
extension SET variable -- for testing purpose).

fix:
- use different keys (200,201,202,203) for the properties of compressed
vectors
- remove the -optional- vector_type serialization for flat vectors
(because it then would repeat for constant and dictionary vectors)

In addition, I enabled the use of compressed vectors in composed types
array, list and struct.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6eb21b1dfe2724ace2a1fde051dd704e7665e31e