[MultiFileReader] Create "local" filters to hand to underlying readers (https://redirect.github.com/duckdb/duckdb/pull/16838)
This PR is a follow up to https://redirect.github.com/duckdb/duckdb/pull/16630 

The aim of this PR is to not require the underlying reader to make use
of complicated mappings created in the `MultiFileReaderData` to
correctly handle filters pushed into the reader.

Previously the `TableFilterSet` handed to the underlying reader
contained "global" filters, i.e the `column_ids` referenced the "global"
schema, which is not guaranteed to match the schema of the file that the
reader is reading.

To convert these into the "local" schema (that of the file that is being
read), the `filter_map` had to be used, then the filter entry had to be
checked to see if the filter points to a column that is present in the
file being read, if it's not then the filter points to the entry in the
`constant_map` that should be used instead.

----------------------------------

After this PR, the underlying reader is handed "local" filters, that
already contain `column_ids` that reference the "local" schema.
No filter will be found that references columns that are not present in
the file, as they have been evaluated beforehand.

These filters (the ones targeting columns that are not in the file) can
be safely removed because if the filter doesn't match we skip the file
entirely, it does not have to be scanned. If the filter *does* match, it
can be safely ignored by the reader.

----------------------------------

### Open questions / Areas for future work discovered

#### Copied `child_indexes` issue
The `child_indexes` of the global `ColumnIndex` are copied 1-to-1 in the
creation of the mapping still:
```c++
			//! FIXME: local fields are not guaranteed to match with the global fields for this struct
			local_index = ColumnIndex(local_id.GetId(), global_id.GetChildIndexes());
		}
```
As the fixme says, this probably shouldn't happen, because schema
evolution could create a divide between the global and local fields for
a given struct.

#### `DynamicFilter` copying
DynamicFilter works similar to a BoundParameterExpression in that it
contains a shared_ptr to a shared structure to allow the filter to be
updated. This can't really be copied and still function correctly.
Perhaps at this stage the DynamicFilter has served its purpose and can
be safely transformed into a non-dynamic filter?
I did not feel confident in making this decision.

#### `StructFilter` schema evolution
The StructFilter besides a `child_idx` (which suffers from the issue
mentioned above) also has a `child_name`.
Through schema evolution the field could be renamed, this information
should probably be added to the `MultiFileReaderColumnDefinition` to
properly handle renames of struct fields.
Or that information is already present, through the LogicalType attached
to the column definition, but it needs to be used properly.

#### `cast_map` casts
The `cast_map` is still present, and necessary to be respected by the
underlying reader, because delaying this cast would complicate the
filtering logic.

If we look at the following flow:
`MFR <-> READER <-> FILE`

Taking the parquet scanner as an example;
Currently, the `cast_map` is applied in the translation from the `FILE`
to the `READER`, through a CastColumnReader.
This reader isn't anything special, it just applies a DefaultCastAs to
the vector read from the reader it wraps, this can easily be replaced by
a BoundCastExpression that is applied in the `FinalizeChunk`.

The problem with moving this cast up, is that it would be applied in the
translation from `READER` to the `MFR`.
This is a problem because the filter pushdown is done inside the READER
(after the cast is performed), and so the filters are using (and
expecting) the "global" type, the target of the cast.

To move the cast up into the translation from `READER` to the `MFR` the
filters that are sent to the reader would need to be changed to work on
the "local" type instead, which is not an easy refactor.

Upstream-commit: https://github.com/duckdb/duckdb/commit/edeb947e2a5939bf45acbdd6be00793b31300546