Python -  Arrow IPC support in from_arrow (https://redirect.github.com/duckdb/duckdb/pull/16821)
This PR implements support in the `from_arrow` function to the
`MessageReader` so we can read IPC Buffers directly from Python - Arrow,
without the need of the users to serialize buffers and send pointers
themselves.

This functionality depends on the `nanoarrow` community extension. The
core-tests are actually in that extension (See
https://github.com/paleolimbot/duckdb-nanoarrow/pull/15 ), while the
only test added here is to check if the function is called on
`MessageReader`, an error message is then printed to let the user know
the extension is missing.

Upstream-commit: https://github.com/duckdb/duckdb/commit/71a5a11750c397114e343ba1670dd8bc8b4266fe