[ART] Introduce a new ARTScanner and make InitMerge and Vacuum iterative (https://redirect.github.com/duckdb/duckdb/pull/16861)
Most ART algorithms use a recursive approach, which limits the ART key
size, can cause long stacks, and other issues. This PR introduces a new
`ARTScanner`, which (for now) implements an iterative strategy for
`InitMerge` and `Vacuum`.

The current `Scan` function of the `ARTScanner` is fairly minimal and
focuses on a fast full-traversal of the index.

1. It handles `Node` types, not their underlying specific types.
2. It only once references an underlying `Node` type when iterating all
its children, allowing it to emplaces all children at once.
3. Handling happens on `Emplace` (`Vacuum`) or `Pop` (`InitMerge`),
depending on the `ARTScanHandlingMode`.
4. For `EMPLACE`, we use the `ARTScanResult` to decide if a child is
pushed to the stack. This is impossible for `POP` handling, where all
children are pushed to the stack.
5. We use the `exhausted` flag for the `Pop` handling mode, as we cannot
alter a node before we've altered all its children, but we also have to
remember that we've processed it.

We iterate on the top element on the stack. The traversal strategy is to
push all children to the stack from left to right for each non-leaf node
(if not `ARTScanHandlingMode::EMPLACE` + `ARTScanResult::SKIP`). We mark
a node as exhausted (without popping it), once all its children are on
the stack (or, if there are no children, immediately).

We should be able to also use this `Scan` function (maybe with some
slight alterations) for the below operations, but I did not want to make
the PR too involved.
- `TransformToDeprecated`
- `VerifyAllocations`

### Next steps

In a follow-up PR, I envision the ARTScanner implementing a second scan
function, which focuses on fast depth-first search and traversal. This
PR's scan quickly processes *all* nodes of the ART. Instead, this second
scan focuses on depth-based operations, such as `Insert`. That scan
function must be more involved, as it needs to remember the currently
active byte of a node, the current key bytes, whether the scanner passed
a gate node, etc. However, we should be able to cover almost all
remaining ART operations with that second scan function.

Finally, we'll likely need a third scan function to merge ARTs, as we'll
need to traverse two ARTs simultaneously. I am keeping that one for
last, as the other scan functions will likely have matured more, and we
can hopefully reuse most (or all) of their functionality.

### Future future work
- Arena-allocate the stack of the scanner and other cool things.

Upstream-commit: https://github.com/duckdb/duckdb/commit/1253553f5fc178f570be6bc0bc3bd6ac0186e3f9