Reduce allocations during aggregations (https://redirect.github.com/duckdb/duckdb/pull/16849)
This PR started off as a minor performance improvement to the string
hashing function (to use a fixed `Load` instead of a variable-size
`memcpy`) but ended up with a lot of effort to reduce allocations during
aggregation, which I profiled using `heaptrack`.

In the current `main`, when doing `SELECT count(DISTINCT UserID) FROM
hits;` (ClickBench), DuckDB would do >700K allocations (this includes
startup, which does a bunch of allocations, e.g.,
`SingleFileBlockManager::LoadFreeList()` and
`InMemoryLogStorage::InMemoryLogStorage`. Note that this is on
`c6a.metal` (192 vCPUs).

With the changes in this PR, the total allocation count is down to just
>300K. Considering that this includes the startup, the total number of
allocations done during aggregation has gone down significantly.

Note that this does not reduce RSS (it's about the same), but just the
number of allocations, which should reduce contention as memory is a
shared resource.

EDIT: I checked how many allocations there are if we don't do a grouped
aggregation, i.e. `select sum(userid) from hits;`, and it's ~141k. So
initially, the query did ~708k allocations, of which 708 - 141 = ~567k
were done by the grouped aggregation. Now, we're doing 316k allocations,
of which 316 - 141 = ~175k are done by the grouped aggregation, which is
a >3x reduction.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b53879b1a84aedb28a72513e03f9c41454e6d094