Fix USING KEY reference error (https://redirect.github.com/duckdb/duckdb/pull/16906)
This PR fixes a bug found in the nightly CI tests:
```sql
WITH RECURSIVE tbl(a, b) USING KEY (a) AS MATERIALIZED (
  SELECT 5, 1
    UNION
  SELECT a, b + 1
  FROM tbl WHERE b < a),
tbl1(a,b) AS (SELECT * FROM recurring.tbl)
SELECT * FROM tbl1;
```

The CTE `tbl1` should not be able to reference `recurring.tbl`, the
Binder should detect that and throw an appropriate error. This, however,
did not work as expected when the `USING KEY` CTE was `MATERIALIZED`.
Then, a `shared_pointer` dereferenced to `NULL`. This is now fixed.

Upstream-commit: https://github.com/duckdb/duckdb/commit/daf2fa367455fc83bd8ba12ee76c0315e232d980