MultiFileReader - Perform nested remapping of field indexes instead of relying on casts (https://redirect.github.com/duckdb/duckdb/pull/16941)
This PR reworks the MultiFileColumnMapper to perform nested remapping of
field indexes instead of relying on casts for mapping of structs. This
allows field identifiers to correctly track schema evolution also across
nested types.

In order to facilitate this, we also introduce a new method -
`remap_struct` - that allows remapping of nested structs across a given
set of keys together with a set of substitutions for keys that are not
present. The `remap_struct` method looks like this:

```sql
SELECT remap_struct(
    {'i': 1, 'j': 2},                    -- input struct
    NULL::ROW(v1 INT, v2 INT, v3 INT),   -- target type
    {'v1': 'j', 'v3': 'i'},              -- remapping keys of input struct
    {'v2': NULL::INTEGER}                -- default value substitutions for values not present in input struct
) AS res;
┌────────────────────────────────────────────┐
│                    res                     │
│ struct(v1 integer, v2 integer, v3 integer) │
├────────────────────────────────────────────┤
│ {'v1': 2, 'v2': NULL, 'v3': 1}             │
└────────────────────────────────────────────┘
```

The new remapping is not only used for field ids but also for remapping
by name. As a result of this, we can now support projection pushdown and
filter pushdown across files that have different struct schemas. The
included benchmark
`benchmark/tpch/parquet/parquet_mixed_struct_projection.benchmark` is an
example of this.

Running the below query on Parquet files with mixed schemas on TPC-H
SF10 gives the following performance:


```sql
SELECT MAX(lineitem.l_linenumber) FROM 'lineitem_mixed_struct*.parquet';
```

| v1.2  |  New  |
|-------|-------|
| 0.72s | 0.07s |

Upstream-commit: https://github.com/duckdb/duckdb/commit/b75d9c37526ad288ab4909c8e848cda85d638c5c